refactor: Move randomCode() to its own utils file
