+++ src/aux-records/AuthUtils.ts
@@ -1,5 +1,3 @@
-import { padStart } from 'lodash';
-import { randomBytes } from 'tweetnacl';
 import {
     fromBase64String,
     toBase64String,
@@ -11,23 +9,7 @@ import { fromByteArray, toByteArray } from 'base64-js';
 import type { UserRole } from './AuthStore';
 
 /**
- * The number of characters that random codes should contain.
  */
-export const RANDOM_CODE_LENGTH = 6;
-
-/**
- * Creates a new random numerical code.
- */
-export function randomCode(): string {
-    const bytes = randomBytes(4);
-    const int32 = new Uint32Array(bytes.buffer);
-    const str = padStart(
-        int32[0].toString().substring(0, RANDOM_CODE_LENGTH),
-        RANDOM_CODE_LENGTH,
-        '0'
-    );
-    return str;
-}
 
 /**
  * Formats the given user ID, session ID, session secret, and expiration time into a key that is used to authenticate a user to a particular session.
